backend/requirements.txt
accelerate==1.7.0
annotated-types==0.7.0
anyio==4.9.0
av==14.4.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.0
diffusers==0.33.1
exceptiongroup==1.3.0
fastapi==0.115.12
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
imagine-sdk @ file:///C:/Users/Qualcomm/Desktop/test_repo_api/imagine_sdk-0.4.2-py3-none-any.whl
importlib_metadata==8.7.0
Jinja2==3.1.6
Markdown==3.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
onnxruntime==1.22.0
opencv-python==4.11.0.86
orjson==3.10.18
packaging==25.0
pillow==11.2.1
protobuf==6.31.1
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.26.1
pyreadline3==3.5.4
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
zipp==3.23.0